Recreate (insecure) archistar-core/bft in Akka.io
This is a prototype implementation of a BFT-enabled distributed
storage system in the spirit of archistar-core/bft. The overall
archistar architecture seemed a good fit for the Actor-pattern so
let's try to find synergies.

Currently a very simplistic test-case works. It is insecure as I'm
not using any encryption or authentication at all. In addition
finished transcations are never freed, so the overall system will
go out-of-memory.

We'll focus on the following features in the near future:

* create good testcases with TestKitJava's aid
* use DeathWatch for Transaction management
* add secure (P)BFT security (macs and encryption)
* try to use Supervision hierarchies to model Eras/Views
* employ remoting to create a ``real'' network interface
* implement a client-side S3 interface
* investigate functional Java8 more (or switch to Scala)
* decouple Transactions from Replicas. Maybe we can create
  direct Transcation<>Transaction communication to improve
  parallelism
* embrace the 'let it crash' paradigm