move transaction generation into View
Before that the Replica waited to get <ClientCommand, Preprepare> before
it told the view to create a new transaction. Now it just forwards messages
to the view and lets the view handle transaction creation on its own.

This was needed as it could be possible that a non-primary replica would
receive knowledge of a new transaction via a Prepare message (if it was
slower than the other replicas).

Also introduce ReplicaMessage as superclass of (Prepreare, Prepare and
Commit) to make message handling easier.